user-service/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
prometheus-flask-exporter==0.23.0
prometheus_client==0.20.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-engineio==4.9.1
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
wsproto==1.2.0
